requirements.txt
# API and HTTP requests
requests>=2.31.0

# Database connection
psycopg[binary]>=3.1.0

# Environment configuration
python-dotenv>=1.0.0

# Retry logic for API calls
tenacity>=8.2.0 

# Apache Spark
pyspark>=3.4.0

# ClickHouse connection
clickhouse-connect>=0.6.0